levantamento-normativos/requirements.txt
streamlit>=1.33.0
requests
openpyxl
beautifulsoup4
google-generativeai
googlesearch-python
